Rename qcm-image to qcm-graphic
