+++ packages/@coorpacademy-app-player/src/view/state-to-props/test/fixtures/slides/qcm-graphic.json
