+++ packages/@coorpacademy-components/src/molecule/answer/index.js
@@ -5,7 +5,7 @@ import DropDown from '../questions/drop-down';
 import FreeText from '../questions/free-text';
 import Picker from '../questions/picker';
 import Qcm from '../questions/qcm';
-import QcmImage from '../questions/qcm-image';
+import QcmGraphic from '../questions/qcm-graphic';
 import QuestionRange from '../questions/question-range';
 import Template from '../questions/template';
 import style from './style.css';
@@ -30,7 +30,7 @@ const Answer = props => {
       case 'qcm':
         return <Qcm {...model} />;
       case 'qcmGraphic':
-        return <QcmImage {...model} />;
+        return <QcmGraphic {...model} />;
       case 'freeText':
         return <FreeText {...model} />;
       case 'dropDown':
