+++ packages/@coorpacademy-components/src/molecule/answer/test/fixtures/qcm-graphic.js
@@ -1,0 +1,12 @@
+import QcmGraphic from '../../../questions/qcm-graphic/test/fixtures/default';
+
+const answerProps = QcmGraphic.props;
+
+export default {
+  props: {
+    model: {
+      type: 'qcmGraphic',
+      ...answerProps
+    }
+  }
+};
