--- packages/@coorpacademy-components/src/molecule/answer/test/fixtures/qcm-image.js
@@ -1,12 +1,0 @@
-import QcmImage from '../../../questions/qcm-image/test/fixtures/default';
-
-const answerProps = QcmImage.props;
-
-export default {
-  props: {
-    model: {
-      type: 'qcmGraphic',
-      ...answerProps
-    }
-  }
-};
