+++ packages/@coorpacademy-components/src/molecule/questions/qcm-graphic/index.js
