+++ packages/@coorpacademy-components/src/molecule/questions/qcm-graphic/style.css
