+++ packages/@coorpacademy-components/src/molecule/questions/qcm-graphic/test/fixtures/default.js
