+++ packages/@coorpacademy-components/src/molecule/questions/qcm-graphic/test/fixtures/no-selected.js
