+++ packages/@coorpacademy-components/src/molecule/slides/slides-player/test/fixtures/qcm-graphic.js
@@ -1,4 +1,4 @@
-import Answer from '../../../../answer/test/fixtures/qcm-image';
+import Answer from '../../../../answer/test/fixtures/qcm-graphic';
 import Default from './default';
 
 const answerType = Answer.props;
