+++ packages/@coorpacademy-components/src/template/app-player/player/test/fixtures/qcm-graphic.js
@@ -1,8 +1,8 @@
 import SlidesHeader from '../../../../../molecule/slides/slides-header/test/fixtures/default';
-import QcmImage from '../../../../../molecule/slides/slides-player/test/fixtures/qcm-image';
+import QcmGraphic from '../../../../../molecule/slides/slides-player/test/fixtures/qcm-graphic';
 
 const headerProps = SlidesHeader.props;
-const playerProps = QcmImage.props;
+const playerProps = QcmGraphic.props;
 
 export default {
   props: {
